changes to index.tsx
